refactor(cosmic-swingset): Move processAction From block-manager into launch-chain
